Embed the .wasm module as base64 encoded string (#115)
* Embed the .wasm module as base64 encoded string

Encodes the wasm module as a base64 encoded string and embeds it in a
loader js file that preloads it on import.

This way the library is compatible with the browser and nodejs
environments out of the box.

Additionally any bundling or lazy evaluation logic layered on top of it
will have no issues, since there are no other dependencies apart from 2
JS files.

* Custom build script

* bump version